CoreWindow: move to UI.Controls, make caption buttons not focusable, and unit test
