+++ FluentAvalonia/Core/ApplicationModel/CoreApplicationViewTitleBar.cs
@@ -1,4 +1,5 @@
 ﻿using FluentAvalonia.Interop;
+using FluentAvalonia.UI.Controls;
 
 namespace FluentAvalonia.Core.ApplicationModel
 {
