+++ FluentAvalonia/Core/ApplicationModel/ICoreApplicationView.cs
@@ -1,0 +1,7 @@
+﻿namespace FluentAvalonia.Core.ApplicationModel
+{
+	public interface ICoreApplicationView
+	{
+		CoreApplicationViewTitleBar TitleBar { get; }
+	}
+}
